+++ lib/src/rule.dart
@@ -10,10 +10,11 @@ class Rule {
     if (unsupportedAttribute.contains(attribute)) {
       attribute = '';
     }
-    if (value.startsWith('function:')) {
+    if (value.startsWith('function:') || value.startsWith('dart.')) {
       attribute = '';
       protocol = 'function';
       function = value.replaceAll('function:', '');
+      function = value.replaceAll('dart.', '');
     } else if (value.startsWith(r'$')) {
       protocol = 'jsonpath';
       rule = value;